scripts/requirements.txt
markdown==3.4.3
PyYAML==6.0.1
python-frontmatter==1.0.0
watchdog==2.1.9
Jinja2==3.1.3
beautifulsoup4==4.12.3
